use worker and distributor endpoint templates
