+++ src/NServiceBus.Msmq.AcceptanceTests/Distributor/When_worker_uses_delayed_retries.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Threading.Tasks;
     using AcceptanceTesting;
-    using Features;
     using NServiceBus.Routing.Legacy;
     using NUnit.Framework;
 
@@ -32,10 +31,8 @@
             public Worker()
             {
                 var distributorAddress = AcceptanceTesting.Customization.Conventions.EndpointNamingConvention(typeof(Distributor));
-                EndpointSetup<DefaultServer>(c =>
+                EndpointSetup<WorkerEndpointTemplate>(c =>
                 {
-                    c.EnableFeature<TimeoutManager>();
-                    c.EnableFeature<FakeReadyMessageProcessor>();
                     c.EnlistWithLegacyMSMQDistributor(
                         distributorAddress,
                         "ReadyMessages",
@@ -69,11 +66,7 @@
         {
             public Distributor()
             {
-                EndpointSetup<DefaultServer>(c =>
-                {
-                    c.EnableFeature<TimeoutManager>();
-                    c.AddHeaderToAllOutgoingMessages("NServiceBus.Distributor.WorkerSessionId", Guid.NewGuid().ToString());
-                }).AddMapping<DelayedMessage>(typeof(Worker));
+                EndpointSetup<DistributorEndpointTemplate>().AddMapping<DelayedMessage>(typeof(Worker));
             }
 
             class DelayedMessageHandler : IHandleMessages<DelayedMessage>
