+++ src/NServiceBus.Msmq.AcceptanceTests/NServiceBus.Msmq.AcceptanceTests.csproj
@@ -51,11 +51,12 @@
     <Compile Include="ConfigureMsmqTransport.cs" />
     <Compile Include="DefaultPublisher.cs" />
     <Compile Include="DefaultServer.cs" />
+    <Compile Include="Distributor\DistributorEndpointTemplate.cs" />
     <Compile Include="Distributor\DistributorContext.cs" />
-    <Compile Include="Distributor\FakeReadyMessageProcessor.cs" />
     <Compile Include="Distributor\When_replying_to_message_sent_from_worker.cs" />
     <Compile Include="Distributor\When_worker_sends_delayed_message.cs" />
     <Compile Include="Distributor\When_worker_uses_delayed_retries.cs" />
+    <Compile Include="Distributor\WorkerEndpointTemplate.cs" />
     <Compile Include="FakePromotableResourceManager.cs" />
     <Compile Include="NServiceBusAcceptanceTest.cs" />
     <Compile Include="Distributor\ReadyMessageDetector.cs" />
